backend/requirements/dev.txt
# =============================================================================
# Wellfond BMS - Development Requirements
# =============================================================================
# Includes base requirements plus development tools
# =============================================================================

-r base.txt

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest-asyncio==1.3.0
pytest-cov==7.1.0
pytest-xdist==3.8.0
factory-boy==3.3.3
faker==40.5.1

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
black==26.3.1
isort==5.12.0
flake8==6.1.0
mypy==1.20.0
django-stubs==6.0.2

# -----------------------------------------------------------------------------
# Development Tools
# -----------------------------------------------------------------------------
ipdb==0.13.13
ipython==9.10.0
django-extensions==4.1
django-debug-toolbar==6.3.0

# -----------------------------------------------------------------------------
# Documentation
# -----------------------------------------------------------------------------
mkdocs==1.6.1
mkdocs-material==9.6.19

# -----------------------------------------------------------------------------
# Performance Profiling
# -----------------------------------------------------------------------------
django-silk==5.3.0

# -----------------------------------------------------------------------------
# Security Scanning
# -----------------------------------------------------------------------------
bandit==1.8.2
safety==3.4.0
